chore: minor code relocation
